Move more code out of "tinas space" and into "streamliners"
