+++ packages/tinacms/src/toolkit/fields/plugins/mdx-field-plugin/plate/index.tsx
@@ -7,43 +7,14 @@ import { FixedToolbar } from './components/plate-ui/fixed-toolbar';
 import { TooltipProvider } from './components/plate-ui/tooltip';
 import FixedToolbarButtons from './components/fixed-toolbar-buttons';
 import { ToolbarProvider } from './toolbar/toolbar-provider';
-import { Plate } from '@udecode/plate/react';
+import { Plate, PlateEditor } from '@udecode/plate/react';
 import { useCreateEditor } from './hooks/use-create-editor';
 import { editorPlugins } from './plugins/editor-plugins';
 import { FloatingToolbar } from './components/plate-ui/floating-toolbar';
 import FloatingToolbarButtons from './components/floating-toolbar-buttons';
-import {
-  loadAnnotations,
-  getCurrentAnnotations,
-  annotationsFieldToMap,
-  annotationMapToEntries,
-  areAnnotationEntriesEqual,
-} from '../../streamliners/discussion-plugin/utils/tina-integration';
-import type { TinaAnnotationsFieldValue } from '../../streamliners/discussion-plugin/utils/tina-integration';
-import type { AnnotationState } from '../../streamliners/discussion-plugin/utils/annotations-store';
-import { normalizeAnnotations, createEmptyAnnotationState } from '../../streamliners/discussion-plugin/utils/annotations-store';
-import { suggestionPlugin } from '../../streamliners/suggestion-plugin/suggestion-plugin';
-import { useRegisterAnnotationsField } from '../../streamliners/discussion-plugin/utils/final-form-bridge';
-import { annotateSuggestionsWithUserName } from '../../streamliners/suggestion-plugin/utils/annotate-suggestions';
-import { AnnotationSync } from '../../streamliners/discussion-plugin/utils/annotation-sync';
-import { useAnnotationSyncHandler } from '../../streamliners/discussion-plugin/utils/use-annotation-sync';
-import { PlateEditor } from '@udecode/plate/react';
-
-const isAnnotationStateEmpty = (state: AnnotationState) =>
-  Object.keys(state.comments).length === 0 &&
-  Object.keys(state.suggestions).length === 0;
+import { useTinaDiscussion } from '../../streamliners/discussion-plugin/hooks/use-tina-discussion';
 
 export const RichEditor = ({ input, tinaForm, field }: RichTextType) => {
-  // Ensure annotations are registered with Final Form so annotation-only edits mark the form dirty
-  useRegisterAnnotationsField(tinaForm);
-
-  const isReadyForSync = React.useRef(false);
-
-  const annotationsKey = React.useMemo(
-    () => input.name ?? field.name ?? 'rich-text',
-    [input.name, field.name]
-  );
-
   const initialValue = React.useMemo(() => {
     if (field?.parser?.type === 'slatejson') {
       return input.value.children;
@@ -55,38 +26,6 @@ export const RichEditor = ({ input, tinaForm, field }: RichTextType) => {
     }
   }, []);
 
-  const initialAnnotationsRef = React.useRef<AnnotationState | null>(null);
-  const formValues =
-    (tinaForm?.values as Record<string, any> | undefined) ??
-    tinaForm?.finalForm?.getState().values;
-  const annotationsFieldValue =
-    (formValues?.annotations as TinaAnnotationsFieldValue | undefined) ??
-    undefined;
-  const annotationsEntries = Array.isArray(annotationsFieldValue?.entries)
-    ? annotationsFieldValue.entries
-    : undefined;
-
-  const globalAnnotationsMap = React.useMemo<
-    Record<string, AnnotationState>
-  >(() => {
-    return annotationsFieldToMap(annotationsEntries);
-  }, [annotationsEntries]);
-
-  const existingFieldAnnotations = globalAnnotationsMap[annotationsKey];
-  const legacyFieldAnnotations = input.value
-    ?.annotations as AnnotationState | undefined;
-
-  if (!initialAnnotationsRef.current) {
-    const sourceAnnotations =
-      existingFieldAnnotations ?? legacyFieldAnnotations;
-
-    initialAnnotationsRef.current = sourceAnnotations
-      ? normalizeAnnotations(sourceAnnotations)
-      : createEmptyAnnotationState();
-  }
-
-  // No longer need React state - annotations live in plugin options
-
   // Memoize component map so Plate children aren't recreated every render
   const components = React.useMemo(() => Components(), []);
 
@@ -96,105 +35,15 @@ export const RichEditor = ({ input, tinaForm, field }: RichTextType) => {
     components,
   }) as PlateEditor;
 
-  const syncAnnotationsToForm = React.useCallback(() => {
-    if (!tinaForm || !editor) return;
+  // Encapsulate all discussion/suggestion sync logic
+  const { onChange, SyncComponent } = useTinaDiscussion({
+    tinaForm,
+    editor,
+    input,
+    field,
+    normalizeLinksInCodeBlocks,
+  });
 
-    // Get current annotations from plugin options without triggering debug logging
-    const currentState = getCurrentAnnotations(editor);
-
-    const values =
-      (tinaForm.values as Record<string, any> | undefined) ??
-      tinaForm.finalForm?.getState().values;
-    const annotationsValue =
-      (values?.annotations as TinaAnnotationsFieldValue | undefined) ??
-      undefined;
-    const currentEntries = Array.isArray(annotationsValue?.entries)
-      ? annotationsValue.entries
-      : undefined;
-    let annotationsMap = annotationsFieldToMap(currentEntries);
-
-    const currentIsEmpty = isAnnotationStateEmpty(currentState);
-
-    if (!currentIsEmpty || annotationsMap[annotationsKey]) {
-      const nextAnnotations: Record<string, AnnotationState> = {
-        ...annotationsMap,
-      };
-
-      if (currentIsEmpty) {
-        delete nextAnnotations[annotationsKey];
-      } else {
-        nextAnnotations[annotationsKey] = currentState;
-      }
-
-      const nextEntries = annotationMapToEntries(nextAnnotations);
-      const normalizedEntries =
-        nextEntries.length > 0 ? nextEntries : undefined;
-
-      if (!areAnnotationEntriesEqual(currentEntries, normalizedEntries)) {
-        console.log('[RichEditor] syncAnnotationsToForm', {
-          annotationsKey,
-          entryCount: nextEntries.length,
-        });
-        tinaForm.change('annotations.entries' as any, normalizedEntries);
-      }
-    }
-  }, [annotationsKey, tinaForm, editor]);
-
-  const lastEmissionRef = React.useRef<{
-    childrenHash: string;
-    annotationsHash: string;
-  } | null>(null);
-
-  const emitChange = React.useCallback(
-    (children: any[], force: boolean = false) => {
-      if (!editor) return;
-
-      const normalized = children.map(normalizeLinksInCodeBlocks);
-      const userName = editor.getOption(suggestionPlugin, 'currentUserName');
-      const normalizedWithNames = annotateSuggestionsWithUserName(
-        normalized,
-        userName
-      );
-      const childrenHash = JSON.stringify(normalizedWithNames);
-
-      // Get annotations from plugin options
-      const currentAnnotations = getCurrentAnnotations(editor);
-      const annotationsHash = JSON.stringify(currentAnnotations);
-
-      const last = lastEmissionRef.current;
-      if (
-        !force &&
-        last &&
-        last.childrenHash === childrenHash &&
-        last.annotationsHash === annotationsHash
-      ) {
-        return;
-      }
-      lastEmissionRef.current = { childrenHash, annotationsHash };
-      input.onChange({
-        type: 'root',
-        children: normalizedWithNames,
-      });
-      syncAnnotationsToForm();
-    },
-    [input, syncAnnotationsToForm, editor, annotationsKey]
-  );
-
-  // Load annotations into plugin options on mount
-  React.useEffect(() => {
-    if (!editor || !initialAnnotationsRef.current) return;
-    loadAnnotations(editor, initialAnnotationsRef.current);
-    // Ensure we don't trigger the dirty flag too early.
-    setTimeout(() => {
-      isReadyForSync.current = true;
-    }, 0);
-  }, [editor]);
-
-  // Trigger initial emit when editor is ready
-  React.useEffect(() => {
-    if (!editor) return;
-    emitChange(editor.children as any[]);
-  }, [editor, emitChange]);
   // This should be a plugin customization
   const ref = React.useRef<HTMLDivElement>(null);
 
@@ -213,44 +62,33 @@ export const RichEditor = ({ input, tinaForm, field }: RichTextType) => {
       }, 100);
     }
   }, [field.experimental_focusIntent, ref]);
-  //
-  const handleAnnotationSync = useAnnotationSyncHandler({
-    annotationsKey,
-    editor,
-    syncAnnotationsToForm,
-    isReadyForSync,
-  });
 
   return (
     <div ref={ref}>
       <Plate
         editor={editor}
-        onChange={(value) => {
-          emitChange(value.value as any[]);
-        }}
+        onChange={onChange}
       >
-        <AnnotationSync
-          onSync={handleAnnotationSync}
-        />
+        <SyncComponent />
         <EditorContainer>
           <TooltipProvider>
-          <ToolbarProvider
-            tinaForm={tinaForm}
-            templates={field.templates}
-            overrides={
-              field?.toolbarOverride ? field.toolbarOverride : field.overrides
-            }
-          >
-            <FixedToolbar>
-              <FixedToolbarButtons />
-            </FixedToolbar>
-            {field?.overrides?.showFloatingToolbar !== false ? (
-              <FloatingToolbar>
-                <FloatingToolbarButtons />
-              </FloatingToolbar>
-            ) : null}
-          </ToolbarProvider>
-          <Editor />
+            <ToolbarProvider
+              tinaForm={tinaForm}
+              templates={field.templates}
+              overrides={
+                field?.toolbarOverride ? field.toolbarOverride : field.overrides
+              }
+            >
+              <FixedToolbar>
+                <FixedToolbarButtons />
+              </FixedToolbar>
+              {field?.overrides?.showFloatingToolbar !== false ? (
+                <FloatingToolbar>
+                  <FloatingToolbarButtons />
+                </FloatingToolbar>
+              ) : null}
+            </ToolbarProvider>
+            <Editor />
           </TooltipProvider>
         </EditorContainer>
       </Plate>
