+++ packages/tinacms/src/toolkit/fields/plugins/streamliners/discussion-plugin/utils/final-form-bridge.ts
@@ -13,7 +13,6 @@ export const useRegisterAnnotationsField = (
     const form = tinaForm?.finalForm;
     if (!form?.registerField) return;
 
-    // We only need the field registered; no subscriptions required.
     const unregister = form.registerField(fieldPath, () => undefined, {});
     return () => {
       unregister?.();
